+++ src/NuGet.Services.Status.Table/Manual/AddStatusEventManualChangeEntity.cs
@@ -1,0 +1,30 @@
+﻿namespace NuGet.Services.Status.Table.Manual
+{
+    public class AddStatusEventManualChangeEntity : ManualStatusChangeEntity
+    {
+        public AddStatusEventManualChangeEntity(
+            string eventAffectedComponentPath,
+            ComponentStatus eventAffectedComponentStatus,
+            string messageContents,
+            bool eventIsActive)
+            : base(ManualStatusChangeType.AddStatusEvent)
+        {
+            EventAffectedComponentPath = eventAffectedComponentPath;
+            EventAffectedComponentStatus = (int)eventAffectedComponentStatus;
+            MessageContents = messageContents;
+            EventIsActive = eventIsActive;
+        }
+
+        public string EventAffectedComponentPath { get; set; }
+
+        /// <remarks>
+        /// This should be a <see cref="ComponentStatus"/> converted to an enum.
+        /// See https://github.com/Azure/azure-storage-net/issues/383
+        /// </remarks>
+        public int EventAffectedComponentStatus { get; set; }
+
+        public string MessageContents { get; set; }
+
+        public bool EventIsActive { get; set; }
+    }
+}
