+++ src/NuGet.Services.Status.Table/Manual/ManualStatusChangeType.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace NuGet.Services.Status.Table
+namespace NuGet.Services.Status.Table.Manual
 {
     public enum ManualStatusChangeType
     {
